pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neopilot"
version = "0.1.0"
description = "NeoPilot – Agente de IA Co-Pilot Universal para Linux"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "NeoPilot Team" }]
requires-python = ">=3.11"
keywords = ["ai", "agent", "computer-use", "linux", "automation"]

dependencies = [
    # LLM Orchestration
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",

    # Web automation
    "playwright>=1.44.0",

    # Desktop / Input control
    "pyautogui>=0.9.54",
    "python-xlib>=0.33",

    # Accessibility (AT-SPI)
    "pyatspi>=2.46.0",
    "dogtail>=1.0.0",

    # Vision / OCR
    "opencv-python>=4.9.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.3.0",
    "mss>=9.0.1",
    "easyocr>=1.7.1",

    # LibreOffice UNO
    "ooo-dev-tools>=0.47.0",

    # Multi-agent
    "crewai>=0.70.0",
    "pyautogen>=0.4.0",

    # Voice STT
    "faster-whisper>=1.0.3",
    "pyaudio>=0.2.14",
    "webrtcvad>=2.0.10",
    "silero-vad>=5.1.0",

    # Voice TTS
    "piper-tts>=1.2.0",
    "elevenlabs>=1.3.0",
    "sounddevice>=0.4.7",

    # Memory / Vector store
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",

    # Configuration / Validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",

    # Security / Crypto
    "cryptography>=42.0.0",
    "keyring>=25.2.0",

    # UI
    "PyQt6>=6.7.0",
    "rich>=13.7.0",
    "typer>=0.12.0",

    # Utilities
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "tenacity>=8.3.0",
    "schedule>=1.2.2",
    "watchdog>=4.0.0",
    "psutil>=5.9.8",
    "python-slugify>=8.0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
neopilot = "neopilot.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
